+++ packages/components/src/DataTable/stories/State.tsx
@@ -40,6 +40,7 @@ const Template: Story<DataTableProps> = ({ ...args }) => (
 
 export const Loading = Template.bind({})
 Loading.args = {
+  caption: 'DataTable State',
   columns,
   state: 'loading',
 }